o Moved non-test files to utils folder.
